pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "inquirer>=3.4.1",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
]
